Refactor snap details views to own view file
